requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
qrcode==8.2
